+++ extended-ref/coll/gasnet_scratch.h
@@ -35,7 +35,7 @@ struct gasnete_coll_scratch_req_t_ {
 		
   /*this is the sum incoming space of all the peers sending to me*/
   /*for now, for non treeops this is the amount of data that everyone is requesting*/
-  uint64_t incoming_size; 
+  uintptr_t incoming_size;
   
   /*information for all the data for which i am the target*/
   /*for non tree ops these values not used*/
@@ -47,7 +47,7 @@ struct gasnete_coll_scratch_req_t_ {
   /*for non tree ops this information is not used*/
   int num_out_peers; 
   gex_Rank_t *out_peers;
-  uint64_t *out_sizes;
+  uintptr_t *out_sizes;
   
   
 };
